Move `AccountCache` to service package
