+++ android/src/main/kotlin/net/mullvad/mullvadvpn/ui/ServiceConnection.kt
@@ -1,6 +1,5 @@
 package net.mullvad.mullvadvpn.ui
 
-import net.mullvad.mullvadvpn.dataproxy.AccountCache
 import net.mullvad.mullvadvpn.dataproxy.AppVersionInfoCache
 import net.mullvad.mullvadvpn.dataproxy.KeyStatusListener
 import net.mullvad.mullvadvpn.dataproxy.RelayListListener
@@ -8,6 +7,7 @@ import net.mullvad.mullvadvpn.service.ServiceInstance
 
 class ServiceConnection(private val service: ServiceInstance, val mainActivity: MainActivity) {
     val daemon = service.daemon
+    val accountCache = service.accountCache
     val connectionProxy = service.connectionProxy
     val connectivityListener = service.connectivityListener
     val locationInfoCache = service.locationInfoCache
@@ -15,7 +15,6 @@ class ServiceConnection(private val service: ServiceInstance, val mainActivity: 
 
     val keyStatusListener = KeyStatusListener(daemon)
     val appVersionInfoCache = AppVersionInfoCache(mainActivity, daemon, settingsListener)
-    val accountCache = AccountCache(daemon, settingsListener)
     var relayListListener = RelayListListener(daemon, settingsListener)
 
     init {
@@ -24,7 +23,6 @@ class ServiceConnection(private val service: ServiceInstance, val mainActivity: 
     }
 
     fun onDestroy() {
-        accountCache.onDestroy()
         appVersionInfoCache.onDestroy()
         keyStatusListener.onDestroy()
         relayListListener.onDestroy()
